Added generic DB Connection Handler
