+++ package.json
@@ -62,6 +62,7 @@
 		"dotenv": "16.1.4",
 		"fomantic-ui-css": "2.9.2",
 		"fs-extra": "^11.1.1",
+		"graphiql": "^3.1.0",
 		"graphql": "16.6.0",
 		"graphql-http": "1.18.0",
 		"helmet": "^7.1.0",
@@ -135,7 +136,7 @@
 		"html-webpack-plugin": "5.5.1",
 		"htmlencode": "0.0.4",
 		"husky": "4.2.5",
-		"image-size-loader": "0.7.0",
+		"image-size-loader": "^0.7.0",
 		"jest": "29.5.0",
 		"jest-environment-jsdom": "29.5.0",
 		"lint-staged": "11.1.2",
@@ -155,7 +156,7 @@
 		"resolve": "1.22.2",
 		"sass-loader": "13.2.2",
 		"stats-webpack-plugin": "0.7.0",
-		"style-loader": "0.12.3",
+		"style-loader": "3.3.4",
 		"typescript": "4.3.2",
 		"ua-parser-js": "0.7.24",
 		"uglifyjs-webpack-plugin": "2.2.0",
