+++ server/schema/users.js
@@ -1,11 +1,3 @@
-const graphql = require('graphql');
-
-const { getLimitCond } = require('./helper');
-const DBConnection = require('../dbClients/postgresql');
-const { logger } = require('../Logger');
-
-const dbConnection = DBConnection.getInstance();
-
 const {
 	GraphQLObjectType,
 	GraphQLString,
@@ -14,7 +6,17 @@ const {
 	GraphQLList,
 	GraphQLNonNull,
 	GraphQLBoolean,
-} = graphql;
+} = require('graphql');
+
+const { getLimitCond } = require('./helper');
+const GenericDBConnection = require('../dbClients/GenericDBConnection');
+const { logger } = require('../Logger');
+
+let dBInstance;
+
+GenericDBConnection.getInstance('postgresql').then((genericInstance) => {
+	dBInstance = genericInstance.getDBInstance();
+});
 
 const UserType = new GraphQLObjectType({
 	name: 'Users',
@@ -36,7 +38,7 @@ const getUser = {
 			? `WHERE "id" = ${args.id}`
 			: getLimitCond('postgresql', 1);
 		const query = `SELECT "id", "firstName", "lastName", "age" FROM TEST_USERS ${whereClause}`;
-		const rows = await dbConnection.executeQuery(query);
+		const rows = await dBInstance.executeQuery(query);
 		return rows[0];
 	},
 };
@@ -57,7 +59,7 @@ const getUsers = {
 				'WHERE ' + keys.map((key) => `"${key}" = '${args[key]}'`).join(' AND ');
 		}
 		const query = `SELECT "id", "firstName", "lastName", "age" FROM TEST_USERS ${whereClause}`;
-		return await dbConnection.executeQuery(query);
+		return await dBInstance.executeQuery(query);
 	},
 };
 
@@ -72,7 +74,7 @@ const createUser = {
 		try {
 			const { firstName, lastName, age } = args;
 			const query = `INSERT INTO TEST_USERS ("firstName", "lastName", "age") VALUES ('${firstName}', '${lastName}', ${age})`;
-			const result = await dbConnection.executeQuery(query);
+			const result = await dBInstance.executeQuery(query);
 			logger.info(result);
 			return true;
 		} catch (error) {
@@ -93,7 +95,7 @@ const updateUser = {
 		try {
 			const { id, firstName, lastName, age } = args;
 			const query = `UPDATE TEST_USERS SET "firstName" = '${firstName}', "lastName" = '${lastName}', "age" = ${age} WHERE "id" = ${id}`;
-			const result = await dbConnection.executeQuery(query);
+			const result = await dBInstance.executeQuery(query);
 			logger.info(result);
 			return true;
 		} catch (error) {
@@ -110,7 +112,7 @@ const deleteUser = {
 	resolve: async (parent, args) => {
 		try {
 			const query = `DELETE FROM TEST_USERS WHERE "id" = ${args.id}`;
-			const result = await dbConnection.executeQuery(query);
+			const result = await dBInstance.executeQuery(query);
 			logger.info(result);
 			return true;
 		} catch (error) {
